attempt to add search
While Bootstrap 5 may have search enabled through their forms, it turns
out that it's difficult for me to find documentation for this specific
task. Moreover, the pkgdown searchbar never had a button associated with
the search... and for good reason: the search is dynamic; it does not
bring you to a page with your search results; it gives you the search
results as a dropdown menu, so you go to your result by clicking
"enter". It makes sense! I think we can provide a blank page for search
results, but we will need to test out the accessibility for sure.